regexi i ostalo

+++ ISNogometniStadion.Model/IzvjesceGodine.cs
@@ -9,6 +9,6 @@ namespace ISNogometniStadion.Model
         public string Stadion { get; set; }
         public string Grad { get; set; }
         public int BrojProdanihUlaznica { get; set; }
-        public int Zarada { get; set; }
+        public decimal Zarada { get; set; }
     }
 }
